feat: add half-star ratings and new rankings feature
- Implemented half-star ratings for shows and movies, allowing users to click on the left half of a star for half-star ratings and the right half for full-star ratings.
- Added a new "Rank" tab in the sidebar to display rated shows and movies grouped by star rating.
- Introduced a "Tier" view to visualize ratings with poster thumbnails, enabling users to switch between shows and movies.
- Updated data models to support floating-point ratings and adjusted database schema accordingly.
- Enhanced UI components to integrate new rating system and display tier rankings.
- Updated changelog and versioning to reflect the new features in version 0.7.1.

+++ src-tauri/src/commands/movies.rs
@@ -16,7 +16,7 @@ pub struct TrackedMovie {
     pub status: Option<String>,
     pub scheduled_date: Option<String>,
     pub watched: bool,
-    pub rating: Option<i64>,
+    pub rating: Option<f64>,
     pub color: Option<String>,
     pub archived: bool,
 }
@@ -204,7 +204,7 @@ pub async fn get_archived_movies(app: AppHandle) -> Result<Vec<TrackedMovie>, St
 pub async fn update_movie_rating(
     app: AppHandle,
     id: i64,
-    rating: Option<i64>,
+    rating: Option<f64>,
 ) -> Result<(), String> {
     let pool = connection::get_pool(&app).await
         .map_err(|e| format!("Database error: {}", e))?;
@@ -491,7 +491,7 @@ pub struct MovieDetail {
     pub scheduled_date: Option<String>,
     pub watched: bool,
     pub watched_at: Option<String>,
-    pub rating: Option<i64>,
+    pub rating: Option<f64>,
     pub notes: Option<String>,
     pub color: Option<String>,
     pub tags: Option<String>,
